Using state helper and state computed
