+++ addon/-private.js
@@ -1,11 +1,12 @@
-/**
- * Objects created with Object.create(null) don't have a prototype,
- * detect this and spread them to Object. Should be fixed in Microstates.
- */
-export function ensurePrototype(value) {
-  if (value && Object.getPrototypeOf(value) === null) {
-    return { ...value };
+import { create, from } from "./index";
+
+export function stateFrom(typeOrValue, value) {
+  if (typeOrValue && typeOrValue.constructor && typeOrValue.constructor.Type) {
+    // already a microstate
+    return typeOrValue;
+  } else if (typeof typeOrValue === "function") {
+    return create(typeOrValue, value);
   } else {
-    return value;
+    return from(typeOrValue);
   }
 }
