+++ addon/macros/state.js
@@ -1,0 +1,16 @@
+import { computed } from "@ember/object";
+import { Store } from "../index";
+
+import { stateFrom } from "../-private";
+
+export default function use(typeOrValue, value) {
+  return computed({
+    get(key) {
+      let initial = stateFrom(typeOrValue, value);
+      return Store(initial, state => this.set(key, state));
+    },
+    set(key, state) {
+      return state;
+    }
+  });
+}
