+++ app/helpers/state.js
@@ -1,0 +1,1 @@
+export { default, state } from 'ember-microstates/helpers/state';
