--- app/helpers/use-state.js
@@ -1,1 +1,0 @@
-export { default, useState } from 'ember-microstates/helpers/use-state';
