--- app/helpers/use-type.js
@@ -1,1 +1,0 @@
-export { default, useType } from 'ember-microstates/helpers/use-type';
