--- app/helpers/use.js
@@ -1,1 +1,0 @@
-export { default, use } from 'ember-microstates/helpers/use';
